pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "real-estate-prediction"
version = "1.0.0"
description = "AI-powered real estate price prediction API with FastAPI and React"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Real Estate Intelligence Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "joblib",
    "pyyaml",
    "python-dotenv",
    "matplotlib",
    "seaborn",
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "jupyter",
    "notebook",
    "ipykernel",
    "pytest",
    "pytest-cov",
    "python-multipart",
    "requests",
    "redis",
    "slowapi",
    "prometheus-client",
    "sentry-sdk[fastapi]",
    "structlog",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pre-commit",
    "mlflow",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
]

[project.urls]
Homepage = "https://github.com/real-estate/prediction"
Repository = "https://github.com/real-estate/prediction"
Documentation = "https://github.com/real-estate/prediction#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "api*"]
exclude = ["tests*", "venv*", "frontend*"]

# =============================================================================
# RUFF CONFIGURATION - Linting and Formatting
# =============================================================================
[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "D",    # pydocstyle
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "PL",   # Pylint
]
ignore = [
    "D100",  # Missing docstring in public module
    "D104",  # Missing docstring in public package
    "D107",  # Missing docstring in __init__
    "D203",  # 1 blank line required before class docstring
    "D213",  # Multi-line docstring summary should start at the second line
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.isort]
known-first-party = ["src", "api"]

# =============================================================================
# MYPY CONFIGURATION - Type Checking
# =============================================================================
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "sklearn.*",
    "matplotlib.*",
    "seaborn.*",
    "joblib.*",
]
ignore_missing_imports = true

# =============================================================================
# PYTEST CONFIGURATION
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--cov=src",
    "--cov=api",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# =============================================================================
# COVERAGE CONFIGURATION
# =============================================================================
[tool.coverage.run]
source = ["src", "api"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/frontend/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = false
fail_under = 80

[tool.coverage.html]
directory = "htmlcov"

# =============================================================================
# BLACK CONFIGURATION (fallback if not using ruff)
# =============================================================================
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | frontend/node_modules
)/
'''
